… ggml/src/iqk/iqk_quantize.cpp
… ggml/src/iqk/iqk_quantize.h
… src/llama.cpp